requirements.txt
Django==1.6.2
Fabric==1.4.3
beautifulsoup4==4.3.2
dj-database-url==0.2.2
dj-static==0.0.5
gdata==2.0.17
geopy==0.95.1
greenlet==0.4.0
gunicorn==0.14.6
psycopg2==2.3.2
pycrypto==2.3
pystache==0.5.3
python-dateutil==2.1
python-memcached==1.53
six==1.5.2
ssh==1.8.0
static==1.0.2
wsgiref==0.1.2
xlrd==0.9.2
